+++ package.json
@@ -78,6 +78,7 @@
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
     "react-redux": "^5.1.1",
+    "react-sortable-hoc": "^1.8.3",
     "react-testing-library": "^6.0.0",
     "redux": "^4.0.1",
     "redux-mock-store": "^1.5.3",
