+++ static/sass/_snapcraft_market_screenshots.scss
@@ -2,62 +2,56 @@
   .p-listing-images {
     min-height: 120px;
     position: relative;
+  }
 
-    .p-listing-images__image {
-      align-items: center;
-      background: $color-light;
-      border: 1px solid $color-light;
-      border-radius: $border-radius;
-      color: $color-link;
-      display: flex;
-      flex-grow: 0;
-      height: 120px;
-      justify-content: center;
-      margin-bottom: $spv-intra;
-      min-width: auto;
-      position: relative;
-      width: 120px;
+  .p-listing-images__image {
+    align-items: center;
+    background: $color-light;
+    border: 1px solid $color-light;
+    color: $color-link;
+    cursor: grab;
+    display: flex;
+    flex-grow: 0;
+    height: 120px;
+    justify-content: center;
+    margin-bottom: $spv-intra;
+    min-width: auto;
+    position: relative;
+    width: 120px;
 
-      .p-listing-images__delete-image {
-        background: $color-x-light;
-        box-sizing: content-box;
-        cursor: pointer;
-        height: 1rem;
-        opacity: .5;
-        padding: .25rem;
-        position: absolute;
-        right: 0;
-        top: 0;
-        width: 1rem;
-        z-index: 1;
-      }
+    .p-listing-images__delete-image {
+      background: $color-x-light;
+      box-sizing: content-box;
+      cursor: pointer;
+      height: 1rem;
+      opacity: .5;
+      padding: .25rem;
+      position: absolute;
+      right: 0;
+      top: 0;
+      width: 1rem;
+      z-index: 1;
+    }
 
-      .p-listing-images__delete-image .p-icon--delete {
-        display: block;
-        top: 0;
-      }
+    .p-listing-images__delete-image .p-icon--delete {
+      display: block;
+      top: 0;
+    }
 
-      &--no-show {
-        justify-self: flex-start;
-
-        img {
-          filter: sepia(100%) hue-rotate(300deg);
-        }
-      }
-
-      &:not(.is-empty):hover,
-      &:not(.is-empty):focus {
-        border: 1px solid $color-link;
-
-        .p-listing-images__delete-image {
-          opacity: 1;
-        }
-      }
+    &--no-show {
+      justify-self: flex-start;
 
       img {
-        align-self: center;
-        max-height: 100%;
-        max-width: 100%;
+        filter: sepia(100%) hue-rotate(300deg);
+      }
+    }
+
+    &:not(.is-empty):hover,
+    &:not(.is-empty):focus {
+      border: 1px solid $color-link;
+
+      .p-listing-images__delete-image {
+        opacity: 1;
       }
     }
 
@@ -68,5 +62,19 @@
       position: static;
       z-index: 1;
     }
+
+    img {
+      align-self: center;
+      max-height: 100%;
+      max-width: 100%;
+      pointer-events: none;
+      user-select: none;
+    }
+  }
+
+  .p-listing-images__add-image {
+    background: $color-x-light;
+    border: 1px solid $color-mid-light;
+    cursor: pointer;
   }
 }
